docs: update install instructions
Quite a few changes were made to the way in which one installs grizli,
most significant of which was to clean up `setup.py` and move to a
`pyproject.toml` + `setup.cfg` method of installation.

This commit updates the installation instructions to reflect these
changes.

Additional changes are as follows:

 - Blocks of code have been all broken down and shifted into
 `code-block` roles, to aid readability.

 - Restructured the different sections to make the document flow better,
 with proper separation of instructions for initial setup, configuration
 file setup and additional package installation.